# lib/api-zod/src/generated/types/event.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Receipt-to-Wallet API
 * OpenAPI spec version: 0.1.0
 */
import type { EventEventType } from './eventEventType';
import type { EventMetadata } from './eventMetadata';

/**
 * Core Event Shape, as returned by GET /events.
 */
export interface Event {
  id: string;
  user_id: number;
  event_type: EventEventType;
  /** @nullable */
  category?: string | null;
  occurred_at: string;
  created_at: string;
  amount: number;
  /** @nullable */
  source?: string | null;
  /** @nullable */
  source_id?: string | null;
  hidden?: boolean;
  /** @nullable */
  status?: string | null;
  metadata?: EventMetadata;
}