# lib/api-zod/src/generated/types/eventLogRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Receipt-to-Wallet API
 * OpenAPI spec version: 0.1.0
 */
import type { EventLogRequestEventType } from './eventLogRequestEventType';
import type { EventLogRequestMetadata } from './eventLogRequestMetadata';

/**
 * Core Event Shape, as a request body. `event_type` must be "transaction" -- the only event type this tracker has. See EVENT_CONTRACT_OPENAPI.yaml for the tracker-agnostic version of this shape.
 */
export interface EventLogRequest {
  event_type: EventLogRequestEventType;
  /** Maps to bank_transactions.date, YYYY-MM-DD. */
  occurred_at: string;
  /** The transaction's dollar amount (signed -- positive spending, negative income). */
  amount?: number;
  /**
     * Defaults to "manual" for events logged through this endpoint.
     * @nullable
     */
  source?: string | null;
  /** @nullable */
  source_id?: string | null;
  /**
     * Maps to bank_transactions.userCategory. Valid values come from GET /transactions/categories (defaults + user-created) -- not a fixed enum in this schema since that list is dynamic.
     * @nullable
     */
  category?: string | null;
  /** Maps directly to bank_transactions.ignored. */
  hidden?: boolean;
  /**
     * Not persisted. Finance's real lifecycle field is `pending` (a boolean, surfaced in `metadata.pending` on read) -- deliberately not generalized into a string enum, per EVENT_CONTRACT_SPEC.md's Resolved Decision #3.
     * @nullable
     */
  status?: string | null;
  /** merchantName, merchantNameRaw, currency (all optional). Not a fixed schema -- see EventLogRequest's description. */
  metadata?: EventLogRequestMetadata;
}