# lib/api-zod/src/generated/types/eventsQueryResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Receipt-to-Wallet API
 * OpenAPI spec version: 0.1.0
 */
import type { Event } from './event';

export interface EventsQueryResult {
  events: Event[];
  total: number;
}