# lib/api-zod/src/generated/types/logEventResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Receipt-to-Wallet API
 * OpenAPI spec version: 0.1.0
 */

export interface LogEventResult {
  status: string;
  id: string;
}